Rock Paper Scissors project is complete
Computer random generated picks either rock paper or scissors.
The user choice checks wether one of the choices is either "rock, paper or scissors".
playRound() function compares the choices of the players and add them to a score variable.

Calling playGame() function initializes so that the user can reapeat the process of above.